app/requirements.dev.txt
-r requirements.txt
black==22.12.0
pytest==8.3.3
pytest-django==4.5.2
factory_boy==3.2.1 
pytest-factoryboy==2.5.1
django-stubs==1.13.1
testcontainers==4.8.1
pluggy==1.5.0
pytest-env==1.1.5